refactor: organise models into directories
