+++ SST Alumni/Models/Onboarding/Authentication/AuthenticationState.swift
