+++ SST Alumni/Models/Security Card/LocationValidationState.swift
