+++ SST Alumni/Models/User/MemberType.swift
