+++ SST Alumni/Views/Alumni Services/AlumniServicesRowView.swift
@@ -75,68 +75,3 @@ struct AlumniServicesRowView<Content: View>: View {
         .listRowSeparator(.hidden)
     }
 }
-
-struct ConditionalLink<Content: View, SheetContent: View>: View {
-    
-    var actionURL: URL?
-    var ctaTitle: String?
-    
-    @ViewBuilder
-    var content: (() -> Content)
-    
-    @ViewBuilder
-    var sheetContent: (() -> SheetContent)
-    
-    @State private var isContentSheetPresented = false
-    
-    var body: some View {
-        if let actionURL, let ctaTitle {
-            Link(destination: actionURL) {
-                VStack(spacing: 0) {
-                    content()
-                    
-                    Divider()
-                    
-                    HStack {
-                        Text(ctaTitle)
-                        Spacer()
-                        Image(systemName: "safari")
-                    }
-                    .foregroundStyle(.blue)
-                    .padding()
-                }
-                .background {
-                    RoundedRectangle(cornerRadius: 16)
-                        .fill(Color(uiColor: .systemGray6))
-                }
-            }
-            .buttonStyle(.plain)
-        } else if let ctaTitle {
-            Button {
-                isContentSheetPresented.toggle()
-            } label: {
-                VStack(spacing: 0) {
-                    content()
-                    
-                    Divider()
-                    
-                    HStack {
-                        Text(ctaTitle)
-                        Spacer()
-                        Image(systemName: "chevron.right")
-                    }
-                    .foregroundStyle(.blue)
-                    .padding()
-                }
-                .background {
-                    RoundedRectangle(cornerRadius: 16)
-                        .fill(Color(uiColor: .systemGray6))
-                }
-            }
-            .buttonStyle(.plain)
-            .sheet(isPresented: $isContentSheetPresented) {
-                sheetContent()
-            }
-        }
-    }
-}
